Add files via upload
A1Q1.c - A program that calculates the squares and cubes of the numbers from 0 to 10000
a) The numbers only multiple of 4.50 (only here you may include <math.h>)
b) The numbers only multiple of 5. 

A1Q2.c - A program that prints the following patterns separately, one below the other. Use for loops to generate the patterns. All asterisks (^) should be printed by a single printf statement for one pattern in a loop.

A1Q3.c - A program to compute each employee’s monthly pay. You do not know the number of employees in advance. Each type of employee has a code infront of his designation for example manager has 1.

The salaries are as follows;
(1). Manager: fixed Monthly salary (2) Daily worker: can work 6 days in week and in case on overtime for extra day is 2/3 of daily wages. (3) Hourly worker: can work 50 hours in a week and half of hourly wage for overtime. (4) Commission worker: weekly wages $230 + 5.9% for total sale in a week. (5) labour at product packing: 0.05% of total sum of total item packed in a week.